chore(build): migrate to Go modules and isolate integration tests
