+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/cellofellow/gopiano
+
+go 1.24.5
+
+require golang.org/x/crypto v0.40.0
