+++ gopiano_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package gopiano
 
 import "testing"
@@ -5,35 +7,16 @@ import "testing"
 var client *Client
 
 func Test_Setup(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	client, _ = NewClient(AndroidClient)
 }
 
-func Test_Encrypt_1(t *testing.T) {
-	testString := "foobar"
-	expected := "3c739d4e29b5d6c6"
-	encrypted := client.encrypt(testString)
-	if encrypted != expected {
-		t.Error("encrypt failed.")
-	} else {
-		t.Log("encrypt passed")
-	}
-}
-
-func Test_Decrypt_1(t *testing.T) {
-	expected := "foobar"
-	testString := "95b6027f2d427dc0"
-	decrypted, err := client.decrypt(testString)
-	if err != nil {
-		t.Error(err)
-	}
-	if decrypted != expected {
-		t.Error("decrypt failed.")
-	} else {
-		t.Log("decrypt passed")
-	}
-}
-
 func Test_AuthPartnerLogin_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.AuthPartnerLogin()
 	if err != nil {
 		t.Error(err)
@@ -42,6 +25,9 @@ func Test_AuthPartnerLogin_1(t *testing.T) {
 }
 
 func Test_AuthUserLogin_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.AuthUserLogin("mellowcellofellow@gmail.com", "Great8")
 	if err != nil {
 		t.Error(err)
@@ -50,6 +36,9 @@ func Test_AuthUserLogin_1(t *testing.T) {
 }
 
 func Test_UserCanSubscribe_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.UserCanSubscribe()
 	if err != nil {
 		t.Error(err)
@@ -58,6 +47,9 @@ func Test_UserCanSubscribe_1(t *testing.T) {
 }
 
 func Test_UserBetBookmarks_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.UserGetBookmarks()
 	if err != nil {
 		t.Error(err)
@@ -66,6 +58,9 @@ func Test_UserBetBookmarks_1(t *testing.T) {
 }
 
 func Test_UserGetStationList_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.UserGetStationList(true)
 	if err != nil {
 		t.Error(err)
@@ -74,6 +69,9 @@ func Test_UserGetStationList_1(t *testing.T) {
 }
 
 func Test_UserGetStationListChecksum_1(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping network test in short mode")
+	}
 	response, err := client.UserGetStationListChecksum()
 	if err != nil {
 		t.Error(err)
